Improve Higher or Lower: per-game averages, new data, season fix
- Change best_season_pir from total to average PIR per game (fair across eras)
- Add 3 new medium-tier categories: best season PPG, APG, RPG
- Values now stored as floats (1 decimal for averages)
- Season selector starts at 2007 (first season with stats data)
- Ingest 2007-2009 seasons (rosters, boxscores, aggregated stats)
- Ingest 2000-2006 rosters (usable for TicTacToe/Roster Guess)
- 5+ games/season minimum for per-game average eligibility

+++ backend/app/models/higher_lower.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import Column, DateTime, Float, Integer, String
 
 from app.database import Base
 
@@ -19,8 +19,8 @@ class HigherLowerGame(Base):
     left_player_id = Column(Integer, nullable=True)
     right_player_id = Column(Integer, nullable=True)
     category = Column(String, nullable=True)
-    left_value = Column(Integer, nullable=True)
-    right_value = Column(Integer, nullable=True)
+    left_value = Column(Float, nullable=True)
+    right_value = Column(Float, nullable=True)
     created_at = Column(DateTime, nullable=False, default=datetime.utcnow)
